bump windows-sys
We want access to `windows_sys::Win32::System::Threading::INFINITE`
which doesn't exist on the previous version.

+++ src/lib.rs
@@ -463,7 +463,7 @@ impl Address {
             return None;
         }
 
-        match (*ptr).sa_family as u32 {
+        match (*ptr).sa_family {
             AF_INET => {
                 let ptr: *const SOCKADDR_IN = std::mem::transmute(ptr);
                 let addr: [u8; 4] = ((*ptr).sin_addr.S_un.S_addr).to_ne_bytes();
